% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbert_encode.R
\name{sbert_encode}
\alias{sbert_encode}
\title{ベクトルをSBERTでエンコード}
\usage{
sbert_encode(model, string_vector)
}
\arguments{
\item{model}{SBERTモデルオブジェクト}

\item{string_vector}{ベクトル形式の文字列群}
}
\value{
エンコードされたベクトル群
}
\description{
ベクトルをSBERTでエンコード
}
